// src/types/Job.ts
export interface Job {
  id: number;
  degree: string;
  job_title: string;
  job_country: string;
  job_type: string;
  job_location: string;
  salary_range_from: number;
  salary_range_to: number;
  company_name: string;
  company_location: string;
  xp_lvl: string;
  updated_at: string;
  company_logo: string;
}